unilaterally include Escaped elements to prevent unexpected post-processing matches.
This prevents a sourcepos assertion failing in tasklists when the
leading `[` is escaped.  I think we expect the escaping of it should
prevent parsing as a task list.

This means our AST assertions include "escaped" elements, which I like
the explicitness of, but let's not include them in XML output without
the render option.

This leaves a failing case which I've tried a few approaches to
addressing, none working out so far: "\&trade;" when outputting to
CommonMark drops the "\", because when we're outputting the "&" (in a
lone Escaped element), `nextb` is None.  Buffering the character output
by one (to reliably see the next character to be output) introduces
all kinds of skew issues which make me think I'd be better off writing
gateware.