parse option to leave Escaped in the document.
When neither the parse nor render option is enabled, we then remove
them post-hoc from the tree. We have to do it *after* all Text node
processing is done, so we don't unify an Escaped->Text with a later Text
and post-process it as if it were regular non-Escaped Text.  It's not
too much bookkeeping, and should be reasonably sound*.

We still have the remaining issue of these Escaped elements, when left
in the tree, preventing the CommonMark formatter from seeing that it
needs to escape a given character when the following node is Escaped.

\* famous last words.